+++ src/internal.h
@@ -5,4 +5,4 @@
 extern SDL_Window *G_window;
 extern SDL_Renderer *G_renderer;
 
-extern double fps_info;
+extern int fps_info;
